Refactored DownloadActivity
