+++ ultrasonic/src/main/java/org/moire/ultrasonic/fragment/SelectAlbumFragment.java
@@ -20,6 +20,7 @@ import androidx.fragment.app.Fragment;
 import androidx.navigation.Navigation;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
+import org.jetbrains.annotations.NotNull;
 import org.moire.ultrasonic.R;
 import org.moire.ultrasonic.data.ActiveServerProvider;
 import org.moire.ultrasonic.domain.MusicDirectory;
@@ -411,7 +412,7 @@ public class SelectAlbumFragment extends Fragment {
     }
 
     @Override
-    public void onPrepareOptionsMenu(Menu menu)
+    public void onPrepareOptionsMenu(@NotNull Menu menu)
     {
         super.onPrepareOptionsMenu(menu);
         playAllButton = menu.findItem(R.id.select_album_play_all);
