+++ ultrasonic/src/main/java/org/moire/ultrasonic/util/Util.java
@@ -51,8 +51,7 @@ import android.widget.Toast;
 import androidx.annotation.ColorInt;
 
 import org.moire.ultrasonic.R;
-import org.moire.ultrasonic.activity.DownloadActivity;
-import org.moire.ultrasonic.activity.MainActivity;
+import org.moire.ultrasonic.activity.NavigationActivity;
 import org.moire.ultrasonic.data.ActiveServerProvider;
 import org.moire.ultrasonic.domain.*;
 import org.moire.ultrasonic.domain.MusicDirectory.Entry;
@@ -1020,8 +1019,8 @@ public class Util
 
 	public static void linkButtons(Context context, RemoteViews views, boolean playerActive)
 	{
-
-		Intent intent = new Intent(context, playerActive ? DownloadActivity.class : MainActivity.class);
+		// TODO: Navigate automatically to PlayerFragment if playerActive = true
+		Intent intent = new Intent(context, playerActive ? NavigationActivity.class : NavigationActivity.class);
 		intent.setAction("android.intent.action.MAIN");
 		intent.addCategory("android.intent.category.LAUNCHER");
 		PendingIntent pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_UPDATE_CURRENT);
