+++ ultrasonic/src/main/res/layout/lyrics.xml
@@ -4,53 +4,61 @@
               a:layout_width="fill_parent"
               a:layout_height="fill_parent">
 
-    <include layout="@layout/tab_progress"/>
+    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+        a:id="@+id/lyrics_refresh"
+        a:layout_width="fill_parent"
+        a:layout_height="0dp"
+        a:layout_weight="1.0">
 
-    <ScrollView
-        	a:id="@+id/lyrics_scrollview"
-            a:layout_width="fill_parent"
-            a:layout_height="0dip"
-            a:layout_weight="1.0">
-
-        <LinearLayout
-                a:orientation="vertical"
+        <androidx.core.widget.NestedScrollView
+                a:id="@+id/lyrics_scrollview"
                 a:layout_width="fill_parent"
-                a:layout_height="fill_parent">
-            <TextView
-                    a:id="@+id/lyrics_artist"
-                    a:textAppearance="?android:attr/textAppearanceMedium"
-                    a:gravity="center_horizontal"
+                a:layout_height="wrap_content"
+                a:scrollbars="vertical"
+                a:layout_weight="1.0">
+
+            <LinearLayout
+                    a:orientation="vertical"
                     a:layout_width="fill_parent"
                     a:layout_height="wrap_content"
-                    a:paddingLeft="10dip"
-                    a:paddingRight="10dip"
-                    a:paddingTop="10dip"
-                    a:paddingBottom="4dip"
-                    />
+                    a:paddingVertical="32dip">
+                <TextView
+                        a:id="@+id/lyrics_artist"
+                        a:textAppearance="?android:attr/textAppearanceMedium"
+                        a:gravity="center_horizontal"
+                        a:layout_width="fill_parent"
+                        a:layout_height="wrap_content"
+                        a:paddingLeft="10dip"
+                        a:paddingRight="10dip"
+                        a:paddingTop="10dip"
+                        a:paddingBottom="4dip"
+                        />
 
-            <TextView
-                    a:id="@+id/lyrics_title"
-                    a:textAppearance="?android:attr/textAppearanceMedium"
-                    a:gravity="center_horizontal"
-                    a:layout_width="fill_parent"
-                    a:layout_height="wrap_content"
-                    a:paddingLeft="10dip"
-                    a:paddingRight="10dip"
-                    />
+                <TextView
+                        a:id="@+id/lyrics_title"
+                        a:textAppearance="?android:attr/textAppearanceMedium"
+                        a:gravity="center_horizontal"
+                        a:layout_width="fill_parent"
+                        a:layout_height="wrap_content"
+                        a:paddingLeft="10dip"
+                        a:paddingRight="10dip"
+                        />
 
-            <TextView
-                    a:id="@+id/lyrics_text"
-                    a:textAppearance="?android:attr/textAppearanceSmall"
-                    a:gravity="center_horizontal"
-                    a:layout_width="fill_parent"
-                    a:layout_height="wrap_content"
-                    a:paddingLeft="10dip"
-                    a:paddingRight="10dip"
-                    />
+                <TextView
+                        a:id="@+id/lyrics_text"
+                        a:textAppearance="?android:attr/textAppearanceSmall"
+                        a:gravity="center_horizontal"
+                        a:layout_width="fill_parent"
+                        a:layout_height="wrap_content"
+                        a:paddingTop="32dip"
+                        a:paddingLeft="10dip"
+                        a:paddingRight="10dip"
+                        />
 
-        </LinearLayout>
+            </LinearLayout>
 
-    </ScrollView>
+        </androidx.core.widget.NestedScrollView>
+    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
     <include layout="@layout/now_playing"/>
