+++ ultrasonic/src/main/res/navigation/navigation_graph.xml
@@ -23,63 +23,63 @@
             app:destination="@id/selectAlbumFragment" />
     </fragment>
     <fragment
-        android:id="@+id/downloadFragment"
-        android:name="org.moire.ultrasonic.fragment.DownloadFragment" />
-    <fragment
         android:id="@+id/selectAlbumFragment"
-        android:name="org.moire.ultrasonic.fragment.SelectAlbumFragment"
-        android:label="SelectAlbumFragment" >
+        android:name="org.moire.ultrasonic.fragment.SelectAlbumFragment" >
     </fragment>
     <fragment
         android:id="@+id/searchFragment"
-        android:name="org.moire.ultrasonic.fragment.SearchFragment"
-        android:label="SearchFragment" >
+        android:name="org.moire.ultrasonic.fragment.SearchFragment" >
         <action
             android:id="@+id/searchToSelectAlbum"
             app:destination="@id/selectAlbumFragment" />
     </fragment>
     <fragment
         android:id="@+id/playlistsFragment"
-        android:name="org.moire.ultrasonic.fragment.PlaylistsFragment"
-        android:label="PlaylistsFragment" >
+        android:name="org.moire.ultrasonic.fragment.PlaylistsFragment" >
         <action
             android:id="@+id/playlistsToSelectAlbum"
             app:destination="@id/selectAlbumFragment" />
     </fragment>
     <fragment
         android:id="@+id/sharesFragment"
-        android:name="org.moire.ultrasonic.fragment.SharesFragment"
-        android:label="SharesFragment" >
+        android:name="org.moire.ultrasonic.fragment.SharesFragment" >
         <action
             android:id="@+id/sharesToSelectAlbum"
             app:destination="@id/selectAlbumFragment" />
     </fragment>
     <fragment
         android:id="@+id/bookmarksFragment"
-        android:name="org.moire.ultrasonic.fragment.BookmarksFragment"
-        android:label="BookmarksFragment" />
+        android:name="org.moire.ultrasonic.fragment.BookmarksFragment" />
     <fragment
         android:id="@+id/chatFragment"
-        android:name="org.moire.ultrasonic.fragment.ChatFragment"
-        android:label="ChatFragment" />
+        android:name="org.moire.ultrasonic.fragment.ChatFragment" />
     <fragment
         android:id="@+id/podcastFragment"
-        android:name="org.moire.ultrasonic.fragment.PodcastFragment"
-        android:label="PodcastFragment" >
+        android:name="org.moire.ultrasonic.fragment.PodcastFragment" >
         <action
             android:id="@+id/podcastToSelectAlbum"
             app:destination="@id/selectAlbumFragment" />
     </fragment>
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="org.moire.ultrasonic.fragment.SettingsFragment"
-        android:label="SettingsFragment" />
+        android:name="org.moire.ultrasonic.fragment.SettingsFragment" />
     <fragment
         android:id="@+id/aboutFragment"
-        android:name="org.moire.ultrasonic.fragment.AboutFragment"
-        android:label="AboutFragment" />
+        android:name="org.moire.ultrasonic.fragment.AboutFragment" />
     <fragment
         android:id="@+id/selectGenreFragment"
-        android:name="org.moire.ultrasonic.fragment.SelectGenreFragment"
-        android:label="SelectGenreFragment" />
+        android:name="org.moire.ultrasonic.fragment.SelectGenreFragment" />
+    <fragment
+        android:id="@+id/playerFragment"
+        android:name="org.moire.ultrasonic.fragment.PlayerFragment" >
+        <action
+            android:id="@+id/playerToSelectAlbum"
+            app:destination="@id/selectAlbumFragment" />
+        <action
+            android:id="@+id/playerToLyrics"
+            app:destination="@id/lyricsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/lyricsFragment"
+        android:name="org.moire.ultrasonic.fragment.LyricsFragment" />
 </navigation>
